urbanblocks_pro/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
alembic
geoalchemy2
reportlab
weasyprint
jinja2
passlib[bcrypt]
pandas
shapely
geopandas
pyproj
plotly
joblib
scikit-learn
shap
numpy
tensorflow
torch
